go: indexed message iterator: guard against bad offsets in file (#1234)
### Changelog

- if a file contains offsets that point outside of file bounds, the Go
indexed reader implementation will return `mcap.ErrBadOffset` when
attempting to use them.

### Docs

None.
### Description

There are files being uploaded to Foxglove that have invalid offsets
inside them - these cause errors in `io.Seek`, which can't be told apart
from true IO errors at the moment. This means if a read fails, we can't
tell if it's because of a corrupt file or flaky I/O. This PR lets us
disambiguate by returning a specific error from the library.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->