go: annotate unexpected token errors (#1247)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

Adds in a `UnexpectedTokenError` error class

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Adds in an `UnexpectedTokenError` error class for when the reader
expects a
certain record but something else was retrieved. It could be useful to
know when this error happens to differentiate it from io-related errors.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

invalid header / metadata come through as generic errors

<!--before content goes here-->

</td><td>

<!--after content goes here-->

invalid header / metadata errors can be differentiated

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->