C++: flush underlying writer after every chunk write (#1266)
### Changelog

#### C++
- Added: added a `flush()` method to `IWritable` to flush any buffered
bytes to the underlying writer. This is optional for any external
implementations to implement, and defaults to a no-op.
- Changed: After writing a chunk, the McapWriter will call `flush()` on
the underlying writer. This ensures that any completed chunk does not
get held up in an internal buffer while the next chunk is being
constructed.

### Docs

See new docstring on `flush()`.
### Description

Adds a virtual function to the `IWritable` interface to flush it, with a
default no-op implementation. I don't really know the right policy for
un-chunked files, but my current opinion is that we can leave that to
the caller - they might call flush every few seconds, or every few
messages.

I don't consider this a breaking change. This is almost certainly
ABI-breaking, but we don't support that as a public interface, and I
can't see a way that this change would cause existing code to fail to
compile.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>
<!--before content goes here-->

C++ writers using `FileWriter` might not see a chunk flushed to disk
even after it has been fully finalized
<!--after content goes here-->
</td><td>

C++ writers using `FileWriter` will see a chunk flushed to disk as soon
as it is finished.

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->
Fixes: #1254