rust: buffer chunks to allow writing without Seek (#1274)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- rust: add `use_buffered_chunks` option which, when set, will allow the
mcap writer to write to a writer that doesn't support seeking

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

This change adds in the concept of a `ChunkSink` used to write chunks
to. It can either be `Direct`, where chunks are written directly to the
output, or `Buffered`, where chunks are written to a temporary buffer
before being written to the output.
Since chunks require seeking back to the start of the chunk to write the
length, buffering chunks allows the mcap writer to write to streams that
don't support seeking (like stdout, network responses, etc.), when used
with something like `NoSeek` from binrw.

I spent a while trying to get this to work using generics so that using
`NoSeek` wouldn't be a requirement (and the risk of a runtime error).
Unfortunately the changes started to get quite complicated and would
have resulted in signficant API changes due to how viral the `Write +
Seek` bounds are currently. Ultimately I figured it would be simpler and
more ergonomic to offer a runtime option.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

~There is now a "linear" conformance writer to make sure that this
doesn't regress~ edit: I realise now the conformance tests aren't
working for writing so I've removed. There are "round trip" unit tests
that cover this.

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->
chunks would not be buffered, requiring an output writer that could
`Seek`

</td><td>

<!--after content goes here-->
using `use_buffered_chunks` allows you to use `NoSeek` with a writer
that does not seek

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->