rust: make interfaces Send (#1284)
### Changelog

<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

rust: make sure structs are Send

### Docs

None

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

This PR makes sure that all MCAP readers and writers can be sent between
threads. This is mainly just adding `Send` to the `Decompressor` trait,
but it also adds compile time assertions to make sure we don't regress.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

readers are not `Send`

</td><td>

<!--after content goes here-->

readers are now `Send`

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->