rust: Channel and Schema structs have IDs (#1297)
### Changelog

#### Breaking changes
- adds `id` members to the crate::Channel and crate::Schema structs.
- refactored and separated `Writer::add_channel` into two methods,
`Writer::add_schema` and `Writer::add_channel`.

### Docs
In docstrings, please review if not clear.
### Description

When a message reader recieves schema and channel information, the
schema and channel IDs in the MCAP are hidden from them. This PR adds
those so that readers can observe ID information.

Additionally, this PR updates the Writer interface in these ways:

1. The `add_channel(Channel)` method is replaced with
`add_channel(schema_id, topic, message_encoding, metadata)`. This was
done because the Channel struct now contains a channel ID, and to
continue to use it would force the user to come up with their own
channel ID before calling this function.
2. An `add_schema(...)` method is added to register schemas ahead of
channels. This is needed to allow the caller to register a new schema
without coming up with an ID in advance.
3. The existing `write(Message)` method is kept, but instead of
automatically assigning new IDs for the message channel and schema, it
uses the IDs passed in. This means that when writing functions that
transform MCAP data, the IDs from the input can be kept in the output.

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->