typescript: support chunk indexes with empty message index (#1304)
### Changelog
typescript: support chunk indexes with empty message index

### Docs
None

### Description
When a chunk contains no messages, the corresponding
[ChunkIndex](https://mcap.dev/spec#chunk-index-op0x08)'s message index
is empty. Prior to this PR, the typescript lib threw an exception when a
chunk index with an empty message index was encountered.
This PR changes this to no longer throw when such a chunk index is
encountered. If the message index is empty, the internal method
`#getSortTime` will fallback to the chunk index' start time which is
`0n` and everything works as expected.


https://github.com/foxglove/mcap/blob/e5c9da312727063194e52e84e2b71a7cfaa66f6d/typescript/core/src/ChunkCursor.ts#L228-L229

Note: We could change the `McapIndexedReader` to exclude such chunk
indexes from its internal heap, but I believe that it's not worth adding
additional code for improving performance of such a corner case