Enable most writer conformance tests for Rust (#1313)
### Changelog
The DataEnd CRC calculation and missing writer options have been added
to allow the Rust writer to pass the conformance tests.

### Docs


https://linear.app/foxglove/issue/FG-9567/rust-mcap-writer-doesnt-pass-or-run-conformance-tests

### Description

Previously, the Rust writer could not pass the conformance tests because
it was not possible to control which summary record types would be
included and because it would not compute the CRC for the DataEnd
record.

Now, all current writer conformance tests pass, except the ones using
the "pad" option. This brings the Rust crate up to parity with the C++
library (and possibly others).

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

No writer conformance tests run or pass

</td><td>

All writer conformance tests except those with "pad" pass

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->
Fixes: FG-9567