go: fix lz4 invalid block checksum (#1309)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

cli: fixes error when input mcap file uses lz4 block checksums

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

As noted in #1294 MCAP files created in Rust were failing when run
through the MCAP CLI because of an invalid block checksum check in the
lz4 decoder. This was fixed upstream so this change just updates the
dependency.

I've tested using `mcap doctor` and `mcap cat` on a file with the
correct checksum and it passes with this change and fails without.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->