writer: add options to switch off CRC calculation (#1328)
### Changelog
- Added: rust writeOptions now include `calculate_*_crc[s]` options to
enable or disable CRC calculation.
### Docs

Check docstrings.
### Description

CRC calculation can have a meaningful impact on CPU load and throughput,
and is not appropriate for some use-cases.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->