rust: avoid allocating for message header (#1327)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

fixed: don't allocate buffer for every `Message` in Rust

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Removes the temporary vec used to write out the message header every
message.

I would like to clean up all of these temporary buffers in
`write_record` since the serialized length should be able to be
calculated dynamically. However this is the one that is called most
frequently so it makes sense to do this first.