rust: fix removal of identical channel data (#1339)
### Changelog
- Fixed: fixed a bug where when writing messages using
`mcap::Writer::write(&Message{...})`, introduced in #1297 . Writing
messages on separate channels with identical channel data would cause
channel data for some channels not to be written to the MCAP, resulting
in an invalid MCAP.
### Docs

None.

### Description

MCAP as specified allows a 1:N relationship between channel content and
channel IDs. This is intentional - there may be multiple sources
publishing the exact same topic, and the output MCAP can and should
contain one channel per publisher, not per topic.

#1297 introduced the use of a bimap to track the relationship between
channel / schema IDs and their content. however, I overlooked at the
time a key constraint of the bimap implementation, in that all values on
the left and right side are unique.
This means it could not track multiple identical channels or schemas
with the same ID. This PR introduces an additional (u16, u16) map for
channels and schemas which tracks the mapping of all IDs to their
"canonical" IDs in the bimap. This allows us to track the presence of
multiple records with the same content but different IDs, while also
preserving the quick lookups that the bimap was introduced for.


<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->