rust: rename items in linear reader (#1341)
### Changelog
- Breaking: renamed `mcap::sans_io::read` to
`mcap::sans_io::linear_reader`.
- Breaking: renamed `mcap::sans_io::read::ReadAction` to
`mcap::sans_io::linear_reader::LinearReadEvent`
- Breaking: renamed `LinearReader::next_action()` to
`LinearReader::next_event()`.
- Breaking: renamed `LinearReader::set_written()` to
`LinearReader::notify_read()`.
- Breaking: renamed `mcap::tokio::read::RecordReader` to
`mcap::tokio::linear_reader::LinearReader`
### Docs

Changes are reflected in `cargo doc`.

### Description

This PR revises the names of functions and structs exposed from the
`sans_io` module, because they're badly named as-is and we need to fix
this now before we add more to this module.
Specifically:
1. `ReadAction` gets renamed to `LinearReadEvent`. This is because
there's no definition of "action" in this context that makes sense -
NeedMore is not an action. NeedMore is requesting the caller take some
action, but then GetMessage is not requesting the caller to do anything.
"events" are a term more commonly used in other parsers such as
[SAX](http://www.saxproject.org/event.html), and can be interpreted more
generally.
2. `set_written()` is renamed to `notify_read()`. This better reflects
what it's for (notifying the reader of the result of a read), and I
found the term `written` confusing on a struct meant to be used for
reading.
3. the `tokio::RecordReader` is renamed to LinearReader for consistency.
4. The `sans_io::read` module is renamed to `linear_reader` - This
better reflects its contents. When we add more readers to the sans_io
module, I don't intend to add them all to the `sans_io::read` module, it
doesn't help anyone to nest so deeply.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->