bump minimum go version for dependabot (#1363)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Bumps the minimum go versions so that dependabot can update the mcap cli
version. It has currently been struggling to do this because the
"slices" package is not available in the standard library (it was
released in 1.21).
https://github.com/foxglove/mcap/network/updates/1002878288

This change sets the minimum version to 1.21 to support it. The
workspace version is currently 1.22.5.