cpp: Fix bugs related to message-less chunks (#1375)
### Changelog
- cpp: Fix a few bugs related to message-less chunks

### Docs
None

### Description
PR #1291 added an optimization, which caused the behavior reported in
issue #1355. The intention of that change was to close a non-empty chunk
before writing a message that would overflow the chunk size. It checks
whether the chunk is non-empty _after_ writing `Channel` & `Schema`
records to the chunk. Therefore, it's possible to write a chunk that
contains no messages.

The writer has a small bug in the metadata that it writes into the chunk
index record. The `currentChunkStart_` register is initialized as
`MaxTime`, and only updated when a message is added to the chunk. But
when writing a message-less chunk, we need to write the message start
time as 0.

The indexed message reader has a small bug in the way it detects the
presence of a message index. It only looks for a message index on the
first chunk index record, instead of scanning all chunk index records.
This is why the test case provided in #1355 fails: the first chunk
contains only `Channel` & `Schema` records without messages, and
therefore the first chunk index has no message index.

Fixes: #1355