Fix issue with re-using McapWriter (#1381)
### Changelog

Fix an issue where re-using a writer leads to schemas appearing in mcap
files where they were not used. Documented how to re-use a writer
correctly and added an example.

### Docs

I made changes to the doc comments, how do I deploy those?

### Description

There were some issues in terminate() not clearing all the state it
needs to, as well as in writing the summary where schemas and channels
that don't appear in the file can be written out.

I fixed those issues, added a test for them, added docs for how to
re-use a McapWriter safely, and added an example.

Linear ticket: FG-10773