typescript: support yielding unknown records from stream reader (#1399)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- fix(typescript): include unknown records in stream reader output,
instead of skipping

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

We're currently skipping unknown records in the Typescript
`McapStreamReader`. It would be great to be able to get access to those
unknown records in order to implement the application extensions / user
proposals section of the spec.

This change makes sure those records are yielded instead of skipped.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

`McapStreamReader.nextRecord()` wouldn't yield unknown records
<!--before content goes here-->

</td><td>

Next record will yield unknown records
<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->

Fixes: FG-11986