rust: add record length limit option (#1404)
### Changelog
- rust: add option to raise an error if a record is above a length
limit, rather than potentially allocating too much memory.
### Docs

See docstrings.

### Description

Adds an option the sans-io reader library to limit the length of record
yielded from the reader. This protects the user from corrupt data that
might otherwise cause the reader to try to allocate huge amounts of
address space.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->