python: bump package version (#1406)
### Changelog
None

### Docs

None

### Description

Bump python mcap package version so that the latest changes to the
package can be released to PyPi. There has been a few PRs merged in the
package which haven't found their way into a release yet, and it would
be nice to get those out.

I'm not at all sure that this is the right way to go about doing it, but
mostly wanted to ask for the latest changes to be released.

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->

These are the unreleased fixes:

foxglove/mcap#1316
foxglove/mcap#1340

---------

Co-authored-by: Rikard Wadman <rikard.wadman@volvo.com>
Co-authored-by: Bennett Hardwick <hello@bennett.dev>