rust: re-use chunk buffers in indexed reader (#1412)
### Changelog
### Fixed
- rust: fixed a bug where the index reader would not re-use allocations
for chunk buffers.
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

The indexed reader contains a mechanism to store decompressed chunks in
buffers to be merged in parallel. It includes a `message_count` which is
meant to be used to allow re-use of the buffer once all messages have
been read out of it. However, this was not being decremented, so all
chunks from an input get stored forever until the input is finished.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->