rust: Add option to summary reader to declare file size (#1410)
### Changelog
- Adds a file size check when reading summaries using SummaryReader

### Docs
None

### Description
In most situations where `SummaryReader` is useful, the file size is
known in advance. This can be used to greatly improve bounds checking
when attempting to allocate buffers or seek. While the `SummaryReader`
is somewhat capable of inferring the file size on its own, it does so in
a rather hacky way that's susceptible to misuse.

This change adds an explicit file size to `SummaryReaderOptions`,
further constrains the bounds checking added in #1407, and adds some
unit tests to validate various corner cases.