Update reader.py to remove argparse dependency
reader.py currently uses an unnecessary module and needlessly complicated and confusing syntax to get the path to the .mcap file. I have removed this dependency and made the code significantly more readable and user-friendly by using a simple input() command which tells the user what it wants and waits for an input, rather than using argparse and expecting the end user to understand what parameters should be passed where. The updated version is also significantly more compact.